Agent: add support for `vscode.workspace.applyEdit`
Previously, the agent would crash when we hit on a codepath that used
the `vscode.WorkspaceEdit` class. This class is used to perform more
advanced edits on the workspace (multi-file, create files, rename files,
delete files, ..). This functionality is used by some inline-edit
commands like the `unit-tests` command that writes the generated test
case to the file instead of to the chat.

Now, the agent implements `vscode.WorkspaceEdit` so that it no longer
crashes. There are also several new JSON-RPC endpoints to support
applying workspace edits:

- `workspace/edit`: request sent from server to client to apply an edit across
  multiple files.
- `workspace/didCreateFiles`: notification sent from client to server to
  notify files were created.
- `workspace/didDeleteFiles`: notification sent from client to server to
  notify files were deleted.
- `workspace/didRenameFiles`: notification sent from client to server to
  notify files were renamed.

The current implementation is not perfect. I didn't spend much time
perfecting all the details of how we apply edits. The goal of this PR is
to unblock further progress.